+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 更改日志 | CHANGE LOG
 
+## v1.2.7/2022-08-29
+
+- [v1.2.6 <=> v1.2.7](https:///github.com/Zuoqiu-Yingyi/siyuan-theme-dark-plus/compare/v1.2.5...v1.2.7)
 - 使用 `Less` 重构 `custom-list-guides.css` | Refactored `custom-list-guides.css` with `Less`.
 - 使用 `Less` 重构 `custom-font-family.css` | Refactored `custom-font-family.css` with `Less`.
 - 使用 `Less` 重构 `custom-board-kanban.css` | Refactored `custom-board-kanban.css` with `Less`.